schema and queue
Big change to remove schema so we just have a default now.  Also added event_notification_buffer table to hold events.  This is so we don't look updates while the indexer is down.

+++ sql/func_asset.sql
@@ -2,21 +2,24 @@ CALL drop_functions_by_name('get_asset');
 /
 -- Stored procedure to get all items
 CREATE OR REPLACE FUNCTION get_asset(p_user_id bigint)
-RETURNS TABLE(id BIGINT, asset_nbr character varying, sys_id character varying, fac_code character varying) AS '
+RETURNS TABLE(id BIGINT, fac_code character varying, asset_nbr character varying, sys_id character varying, create_ts timestamptz, update_ts timestamptz) AS '
 BEGIN
     RETURN QUERY
-    SELECT asset.id, asset.asset_nbr, asset.sys_id, facility.fac_code 
-	FROM daas.asset asset
-    JOIN daas.facility facility on asset.fac_id = facility.id
-	JOIN daas.user_facility uf on facility.id = uf.fac_id
-	WHERE uf.user_id = p_user_id;
+    SELECT 
+		asset.id, facility.fac_code , asset.asset_nbr, asset.sys_id, asset.create_ts, asset.update_ts
+	FROM 
+		asset asset
+    	JOIN facility facility on asset.fac_id = facility.id
+		JOIN user_facility uf on facility.id = uf.fac_id
+	WHERE 
+		(uf.user_id = p_user_id OR p_user_id is null);
 END;
 ' LANGUAGE plpgsql;
 /
 CALL drop_functions_by_name('get_asset_by_id');
 /
 -- Stored procedure to get an asset by ID
-CREATE OR REPLACE FUNCTION daas.get_asset_by_id(p_jsonb jsonb, p_user_id bigint)
+CREATE OR REPLACE FUNCTION get_asset_by_id(p_jsonb jsonb, p_user_id bigint)
 RETURNS TABLE(id bigint, fac_code character varying, asset_nbr character varying, sys_id character varying, create_ts timestamptz, update_ts timestamptz)
 AS '
 DECLARE
@@ -24,40 +27,42 @@ BEGIN
   	RETURN QUERY
 	WITH fac_cte_condition as (
 		select f.id, f.fac_code
-		from daas.facility f
+		from facility f
 		JOIN get_jsonb_values_by_key (p_jsonb, ''fac_code'') k on f.fac_code = k.value
 	),
 	fac_cte_no_condition as (
 		select f.id, f.fac_code
-		from daas.facility f
+		from facility f
 	),
 	asset_cte_condition as (
 		select a.id, a.fac_id, a.asset_nbr, a.sys_id, a.create_ts, a.update_ts
-		from daas.asset a
+		from asset a
 		JOIN get_jsonb_values_by_key (p_jsonb, ''asset_nbr'') k on a.asset_nbr = k.value
 	),
 	asset_cte_no_condition as (
 		select a.id, a.fac_id, a.asset_nbr, a.sys_id, a.create_ts, a.update_ts
-		from daas.asset a
+		from asset a
 	)
 
 	select a.id, f.fac_code, a.asset_nbr, a.sys_id, a.create_ts, a.update_ts
 	from
 	(
-		select id, fac_code from fac_cte_condition
+		select fc.id, fc.fac_code from fac_cte_condition fc
 		union all
-		select id, fac_code from fac_cte_no_condition
+		select fnc.id, fnc.fac_code from fac_cte_no_condition fnc
 		where (select count(*) from fac_cte_condition) = 0
 	) f
 	join
 	(
-		select id, fac_id, asset_nbr, sys_id, create_ts, update_ts from asset_cte_condition
+		select ac.id, ac.fac_id, ac.asset_nbr, ac.sys_id, ac.create_ts, ac.update_ts from asset_cte_condition ac
 		union all
-		select id, fac_id, asset_nbr, sys_id, create_ts, update_ts from asset_cte_no_condition
+		select anc.id, anc.fac_id, anc.asset_nbr, anc.sys_id, anc.create_ts, anc.update_ts from asset_cte_no_condition anc
 		where (select count(*) from asset_cte_condition) = 0
 	) a on f.id = a.fac_id
 	join
-	daas.user_facility uf on f.id = uf.fac_id;
+	user_facility uf on f.id = uf.fac_id
+	WHERE
+		(uf.user_id = p_user_id OR p_user_id is null);
 
 END;
 ' LANGUAGE plpgsql;
@@ -81,7 +86,7 @@ BEGIN
 	);
 
     SELECT f.id INTO unknown_fac_id
-    FROM daas.facility_facility f
+    FROM facility_facility f
     WHERE upper(f.fac_code) = ''UNKNOWN'';
 
 	INSERT INTO temp_json_data (asset_nbr, sys_id, fac_code)
@@ -93,16 +98,19 @@ BEGIN
 	FROM jsonb_array_elements(p_jsonb_in::JSONB) AS p_jsonb;
 
 	UPDATE temp_json_data
-	SET fac_id = COALESCE(daas.facility_facility.id, unknown_fac_id)
-	FROM daas.facility_facility
-	WHERE UPPER(temp_json_data.fac_code) = UPPER(daas.facility_facility.fac_code);
+	SET fac_id = COALESCE(facility_facility.id, unknown_fac_id)
+	FROM facility_facility
+	WHERE UPPER(temp_json_data.fac_code) = UPPER(facility_facility.fac_code);
 	
 	-- remove upserts where the user does not have access to the facility
-	DELETE from temp_json_data t
-	WHERE NOT EXISTS (select 1 FROM daas.user_facility uf WHERE t.fac_id = uf.fac_id AND uf.user_id = p_user_id);
+	IF p_user_id IS NOT NULL THEN
+		DELETE from temp_json_data t
+		WHERE 
+			NOT EXISTS (select 1 FROM user_facility uf WHERE t.fac_id = uf.fac_id AND uf.user_id = p_user_id);
+	END IF;
 
 	-- Perform UPSERT: Insert new records or update existing ones
-	MERGE INTO daas.asset AS target
+	MERGE INTO asset AS target
 	USING temp_json_data AS source
 	ON target.asset_nbr = source.asset_nbr
 	WHEN MATCHED THEN
@@ -118,17 +126,19 @@ BEGIN
     -- Raise event for consumers
     FOR asset_nbr IN
         SELECT a.asset_nbr 
-		FROM daas.asset a
+		FROM asset a
 		JOIN temp_json_data t ON a.asset_nbr = t.asset_nbr
     LOOP
+		INSERT INTO event_notification_buffer(channel, payload, create_ts)
+		VALUES (p_channel_name, asset_nbr, now());
         PERFORM pg_notify(p_channel_name, asset_nbr);
     END LOOP;
 	
     -- Return the updated records
     RETURN QUERY 
     SELECT a.id, a.asset_nbr, a.sys_id, f.fac_code
-    FROM daas.asset a
-    JOIN daas.facility_facility f ON a.fac_id = f.id
+    FROM asset a
+    JOIN facility_facility f ON a.fac_id = f.id
 	JOIN temp_json_data t ON a.asset_nbr = t.asset_nbr;
 END;
